Potential fix for code scanning alert no. 37: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in the larger type, and the result will be correctly assigned to start_offset.

The best way to fix this is to cast row to size_t before multiplying it by stride. This change should be made on line 111 of the file libwebp-1.0.3/src/dsp/filters_neon.c.

+++ libwebp-1.0.3/src/dsp/filters_neon.c
@@ -108,7 +108,7 @@ static WEBP_INLINE void DoVerticalFilter_NEON(const uint8_t* in,
                                               int width, int height, int stride,
                                               int row, int num_rows,
                                               uint8_t* out) {
-  const size_t start_offset = row * stride;
+  const size_t start_offset = (size_t)row * stride;
   const int last_row = row + num_rows;
   SANITY_CHECK(in, out);
   in += start_offset;
